Backgrounds work again. Cleanup demos
At some point in getting video feedback to work I thought it might help
using an RGB format for render buffers, but RGBA is fine, allowing the
background command to work as before. Due to being a separate div,
background is not incorporated into the feedback (which could be a
useful way of seeding it).
Some adjustments to syntax in demos as indicated by @davidedc in
comments.
Also using mutate function in doOnce in some demos, to be applied
interactively.